// packages/control-property-editor/src/api.ts
export interface Control {
    /**
     * Unique identifier for a control
     */
    id: string;
    /**
     * control name
     */
    name: string;
    /**
     * control type
     */
    type: string;

    properties: ControlProperty[];
}
export type PropertyValue = string | boolean | number;

export interface PropertyChange<T extends PropertyValue = PropertyValue> {
    controlId: string;
    controlName: string;
    propertyName: string;
    value: T;
}
export interface PropertyChanged<T extends PropertyValue = PropertyValue> {
    controlId: string;
    propertyName: string;
    newValue: T;
}

export interface PropertyChangeFailed {
    controlId: string;
    propertyName: string;
    errorMessage: string;
}
// ...
export type ControlProperty =
    | BooleanControlProperty
    | IntegerControlProperty
    | FloatControlProperty
    | StringControlProperty
    | StringControlPropertyWithOptions;

export interface OutlineNode {
    controlId: string;
    controlType: string; // as used in rta i.e sap.m.NavContainer
    name: string;
    visible: boolean;
    editable: boolean;
    children: OutlineNode[];
}
// ...
export const EXTERNAL_ACTION_PREFIX = '[ext]';

const createExternalAction = createActionFactory(EXTERNAL_ACTION_PREFIX);

export const iconsLoaded = createExternalAction<IconDetails[]>('icons-loaded');
export const controlSelected = createExternalAction<Control>('control-selected');
export const selectControl = createExternalAction<string>('select-control');
export const deletePropertyChanges = createExternalAction<PropertyChangeDeletionDetails>('delete-property-changes');
export const outlineChanged = createExternalAction<OutlineNode[]>('outline-changed');
export const changeProperty = createExternalAction<PropertyChange>('change-property');
export const propertyChanged = createExternalAction<PropertyChanged>('property-changed');
export const propertyChangeFailed = createExternalAction<PropertyChangeFailed>('change-property-failed');